创建标签,查看标签,动态添加标签接口

+++ 邂逅Node/coderhub项目实战/src/service/comment.service.js
@@ -22,9 +22,9 @@ class CommentService {
     ]);
     return result;
   }
-  async removeComment(id) {
-    const statement = "delete from comment where id =?;";
-    const [result] = await connection.execute(statement, [id]);
+  async removeComment(commentId,userId) {
+    const statement = "delete from comment where id = ? and user_id = ?;";
+    const [result] = await connection.execute(statement, [commentId,userId]);
     return result;
   }
 }
